+++ frontend/src/component/content/right/Footer.tsx
