+++ frontend/src/pages/Self.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import styled from 'styled-components';
 import Header from '../component/header/Header';
 import Progress from '../component/progress/Progress';
-import Content from '@/component/Content';
+import Content from '@/component/content/Content';
 
 function Self() {
   return (
