+++ frontend/src/style/theme.tsx
@@ -9,4 +9,5 @@ const colors = {
   Cool_Grey_002: '#DBDEE3',
   Cool_Grey_001: '#F2F4F7',
 };
+
 export {colors};
